+++ tests/test_serde.rs
@@ -88,30 +88,75 @@ async fn time() {
 
     let client = common::prepare_database!();
 
-    #[derive(Debug, Row, Serialize, Deserialize)]
+    #[derive(Debug, PartialEq, Eq, Row, Serialize, Deserialize)]
     struct MyRow {
         #[serde(with = "clickhouse::serde::time::datetime")]
         dt: OffsetDateTime,
+        #[serde(with = "clickhouse::serde::time::datetime64::secs")]
+        dt64s: OffsetDateTime,
+        #[serde(with = "clickhouse::serde::time::datetime64::millis")]
+        dt64ms: OffsetDateTime,
+        #[serde(with = "clickhouse::serde::time::datetime64::micros")]
+        dt64us: OffsetDateTime,
+        #[serde(with = "clickhouse::serde::time::datetime64::nanos")]
+        dt64ns: OffsetDateTime,
+    }
+
+    #[derive(Debug, Row, Deserialize)]
+    struct MyRowStr {
+        dt: String,
+        dt64s: String,
+        dt64ms: String,
+        dt64us: String,
+        dt64ns: String,
     }
 
     client
-        .query("CREATE TABLE test(dt DateTime) ENGINE = MergeTree ORDER BY dt")
+        .query(
+            "
+            CREATE TABLE test(
+                dt DateTime,
+                dt64s DateTime64(0),
+                dt64ms DateTime64(3),
+                dt64us DateTime64(6),
+                dt64ns DateTime64(9)
+            )
+            ENGINE = MergeTree
+            ORDER BY dt
+        ",
+        )
         .execute()
         .await
         .unwrap();
 
-    let dt = datetime!(2022-11-13 15:27:42 UTC);
+    let original_row = MyRow {
+        dt: datetime!(2022-11-13 15:27:42 UTC),
+        dt64s: datetime!(2022-11-13 15:27:42 UTC),
+        dt64ms: datetime!(2022-11-13 15:27:42.123 UTC),
+        dt64us: datetime!(2022-11-13 15:27:42.123456 UTC),
+        dt64ns: datetime!(2022-11-13 15:27:42.123456789 UTC),
+    };
 
     let mut insert = client.insert("test").unwrap();
-    insert.write(&MyRow { dt }).await.unwrap();
+    insert.write(&original_row).await.unwrap();
     insert.end().await.unwrap();
 
-    let (row, row_dt_str) = client
-        .query("SELECT ?fields, toString(dt) FROM test")
-        .fetch_one::<(MyRow, String)>()
+    let row = client
+        .query("SELECT ?fields FROM test")
+        .fetch_one::<MyRow>()
         .await
         .unwrap();
 
-    assert_eq!(row.dt, dt);
-    assert_eq!(row_dt_str, &dt.to_string()[..19]);
+    let row_str = client
+        .query("SELECT * APPLY(toString) FROM test")
+        .fetch_one::<MyRowStr>()
+        .await
+        .unwrap();
+
+    assert_eq!(row, original_row);
+    assert_eq!(row_str.dt, &original_row.dt.to_string()[..19]);
+    assert_eq!(row_str.dt64s, &original_row.dt64s.to_string()[..19]);
+    assert_eq!(row_str.dt64ms, &original_row.dt64ms.to_string()[..23]);
+    assert_eq!(row_str.dt64us, &original_row.dt64us.to_string()[..26]);
+    assert_eq!(row_str.dt64ns, &original_row.dt64ns.to_string()[..29]);
 }